Move config and job_queue
